Move useGetNodeTypeDetail and useGetTimeoutString to hooks directory
