+++ frontend/awx/resources/templates/WorkflowVisualizer/hooks/index.ts
@@ -1,22 +1,24 @@
 import { useCloseSidebar } from './useCloseSidebar';
 import { useCreateEdge } from './useCreateEdge';
 import { useDedupeOldNodes } from './useDedupeOldNodes';
-import { useGetDetailComponent } from './useGetDetailComponent';
+import { useGetNodeTypeDetail } from './useGetNodeTypeDetail';
 import { useRemoveGraphElements } from './useRemoveGraphElements';
 import { useRemoveNode } from './useRemoveNode';
 import { useSaveVisualizer } from './useSaveVisualizer';
 import { useSelectedNode } from './useSelectedNode';
+import { useGetTimeoutString } from './useGetTimeoutString';
 import { useGetInitialValues, useNodeTypeStepDefaults } from './useGetInitialValues';
 
 export {
   useCloseSidebar,
   useCreateEdge,
   useDedupeOldNodes,
-  useGetDetailComponent,
+  useGetNodeTypeDetail,
   useGetInitialValues,
   useNodeTypeStepDefaults,
   useRemoveGraphElements,
   useRemoveNode,
   useSaveVisualizer,
   useSelectedNode,
+  useGetTimeoutString,
 };
